+++ core/vrl/ext/com/jhlabs/image/LightFilter.java
@@ -211,7 +211,7 @@ public class LightFilter extends WholeImageFilter {
 		float width45 = Math.abs(6.0f * bumpHeight);
 		boolean invertBumps = bumpHeight < 0;
 		Vector3f position = new Vector3f(0.0f, 0.0f, 0.0f);
-		Vector3f viewpoint = new Vector3f((float)width / 2.0f, (float)height / 2.0f, viewDistance);
+		Vector3f viewpoint = new Vector3f(width / 2.0f, height / 2.0f, viewDistance);
 		Vector3f normal = new Vector3f();
 		Color4f envColor = new Color4f();
 		Color4f diffuseColor = new Color4f( new Color(material.diffuseColor) );
