+++ core/vrl/ext/com/jhlabs/image/ShatterFilter.java
@@ -133,10 +133,10 @@ public class ShatterFilter extends AbstractBufferedImageOp {
     public BufferedImage filter( BufferedImage src, BufferedImage dst ) {
         if ( dst == null )
             dst = createCompatibleDestImage( src, null );
-        float width = (float)src.getWidth();
-        float height = (float)src.getHeight();
-        float cx = (float)src.getWidth() * centreX;
-        float cy = (float)src.getHeight() * centreY;
+        float width = src.getWidth();
+        float height = src.getHeight();
+        float cx = src.getWidth() * centreX;
+        float cy = src.getHeight() * centreY;
         float imageRadius = (float)Math.sqrt( cx*cx + cy*cy );
 
 //        BufferedImage[] tiles = new BufferedImage[iterations];
