+++ core/vrl/ext/com/jhlabs/image/WarpGrid.java
@@ -16,9 +16,6 @@ limitations under the License.
 
 package vrl.ext.com.jhlabs.image;
 
-import java.awt.*;
-import java.awt.image.*;
-
 /**
  * A warp grid.
  * From "A simplified approach to Image Processing" by Randy Crane
@@ -165,8 +162,8 @@ int j = 0;
 		int index = 0;
 		for (int row = 0; row < rows; row++) {
 			for (int col = 0; col < cols; col++) {
-				intermediate.xGrid[index] = (float)ImageMath.lerp(t, xGrid[index], destination.xGrid[index]);
-				intermediate.yGrid[index] = (float)ImageMath.lerp(t, yGrid[index], destination.yGrid[index]);
+				intermediate.xGrid[index] = ImageMath.lerp(t, xGrid[index], destination.xGrid[index]);
+				intermediate.yGrid[index] = ImageMath.lerp(t, yGrid[index], destination.yGrid[index]);
 				index++;
 			}
 		}
