+++ core/vrl/ext/com/jhlabs/math/Noise.java
@@ -98,7 +98,7 @@ public class Noise implements Function1D, Function2D, Function3D {
 		t = x + N;
 		bx0 = ((int)t) & BM;
 		bx1 = (bx0+1) & BM;
-		rx0 = t - (int)t;
+		rx0 = t - t;
 		rx1 = rx0 - 1.0f;
 
 		sx = sCurve(rx0);
@@ -127,13 +127,13 @@ public class Noise implements Function1D, Function2D, Function3D {
 		t = x + N;
 		bx0 = ((int)t) & BM;
 		bx1 = (bx0+1) & BM;
-		rx0 = t - (int)t;
+		rx0 = t - t;
 		rx1 = rx0 - 1.0f;
 	
 		t = y + N;
 		by0 = ((int)t) & BM;
 		by1 = (by0+1) & BM;
-		ry0 = t - (int)t;
+		ry0 = t - t;
 		ry1 = ry0 - 1.0f;
 	
 		i = p[bx0];
@@ -178,19 +178,19 @@ public class Noise implements Function1D, Function2D, Function3D {
 		t = x + N;
 		bx0 = ((int)t) & BM;
 		bx1 = (bx0+1) & BM;
-		rx0 = t - (int)t;
+		rx0 = t - t;
 		rx1 = rx0 - 1.0f;
 
 		t = y + N;
 		by0 = ((int)t) & BM;
 		by1 = (by0+1) & BM;
-		ry0 = t - (int)t;
+		ry0 = t - t;
 		ry1 = ry0 - 1.0f;
 	
 		t = z + N;
 		bz0 = ((int)t) & BM;
 		bz1 = (bz0+1) & BM;
-		rz0 = t - (int)t;
+		rz0 = t - t;
 		rz1 = rz0 - 1.0f;
 	
 		i = p[bx0];
